Add files via upload
Given the number of elements to be generated in the Fibonacci sequence,n, create a singly linked list
with the number of nodes equal to the last number in the sequence and store alphabets in sequence
in those nodes. Delete the nodes whose indexes are the numbers in the Fibonacci sequence and
display the remaining nodes. If the number of nodes is greater than 26, repeat the alphabets starting
from 'A'